pyproject.toml
[project]
name = "prlens"
version = "0.1.0"
description = "AI-powered Pull Request analyzer using Google Gemini"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.62.0",
    "google-api-core>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "pygithub>=2.8.1",
    "requests>=2.31.0",
    "unidiff>=0.7.5",
    "langgraph>=1.0.8",
]
